requirements.txt
altair==4.1.0
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
eli5==0.11.0
emoji==0.5.4
entrypoints==0.3
flasgger==0.9.4
Flask==1.1.1
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.24
graphviz==0.17
greenlet==1.1.2
gunicorn==19.9.0
idna==3.3
importlib-metadata==4.8.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lightgbm==3.3.1
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numba==0.48.0
numpy==1.20.3
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyparsing==3.0.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scikit-learn==0.22.1
scipy==1.7.1
seaborn==0.10.1
Send2Trash==1.8.0
shap==0.40.0
simplejson==3.17.5
six==1.16.0
slicer==0.0.7
smmap==5.0.0
streamlit==1.1.0
streamlit-aggrid==0.2.2.post4
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
xgboost==1.5.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0